Digeo: r?tablissement de la capacit? ? commencer dans les octaves n?gatives et autre nettoyage de code.

+++ src/uti_phgrm/MICMAC/Jp2ImageLoader.cpp
@@ -233,8 +233,8 @@ void JP2ImageLoader::LoadNCanaux(const std::vector<sLowLevelIm<short unsigned in
 	
 	int * precisions = new int[num_components];
     bool *is_signed = new bool[num_components];
-    for(size_t i=0;i<num_components;++i) precisions[i]=precision;
-    for(size_t i=0;i<num_components;++i) is_signed[i]=signe;
+    for(int i=0;i<num_components;++i) precisions[i]=precision;
+    for(int i=0;i<num_components;++i) is_signed[i]=signe;
 	
     kdu_dims *comp_dims = new kdu_dims[num_components];
     for (n=0; n < num_components; n++)
@@ -388,8 +388,8 @@ void JP2ImageLoader::LoadNCanaux(const std::vector<sLowLevelIm<float> > & aVImag
 	
 	int * precisions = new int[num_components];
     bool *is_signed = new bool[num_components];
-    for(size_t i=0;i<num_components;++i) precisions[i]=precision;
-    for(size_t i=0;i<num_components;++i) is_signed[i]=signe;
+    for(int i=0;i<num_components;++i) precisions[i]=precision;
+    for(int i=0;i<num_components;++i) is_signed[i]=signe;
 	
     kdu_dims *comp_dims = new kdu_dims[num_components];
     for (n=0; n < num_components; n++)
@@ -616,7 +616,6 @@ void JP2ImageLoader::LoadNCanaux(const std::vector<sLowLevelIm<unsigned char> > 
     delete[] comp_dims;
 
 }
-
 
 #endif
